Redo Search
Switched from AVLTree to Java's built in TreeMap, which is an implementation of a red black tree. Even though AVL trees are faster for lookup, the small optimization is not worth the extra size in code, since we also want to minimize disk space. Using a built in type will also hopefully improve readability.